bpf: Protect each uintptr with runtime.KeepAlive
The Go's GC is unable to track objects referred by uintptr [1] [2].

A consequence of this is that the GC might collect such objects
when a blocking syscall is executed. This might corrupt memory of
other Go objects e.g. if the reclaimed memory is used for new objects and
the kernel modifies the objects in the context of the blocking syscall.

To prevent from this, we protect each pointer to objects with
runtime.KeepAlive() which marks the objects as alive at least until
runtime.KeepAlive() has been called.

[1]: https://github.com/golang/go/issues/13372#issuecomment-160655731
[2]: https://utcc.utoronto.ca/~cks/space/blog/programming/GoUintptrVsUnsafePointer

Co-developed-by: André Martins <andre@cilium.io>
Co-developed-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Martynas Pumputis <m@lambda.lt>